+++ src/ops/filesystem.js
@@ -3,6 +3,36 @@ import * as console from "ext:deno_console/01_console.js";
 import * as webidl from "ext:deno_webidl/00_webidl.js";
 import * as fs from "ext:deno_fs/30_fs.js";
 import { WritableStream } from "ext:deno_web/06_streams.js";
+import { SeekMode } from "ext:deno_io/12_io.js";
+import { op_get_location } from "ext:core/ops";
+
+export const errors = {
+  INVALID: ["seeking position failed.", "InvalidStateError"],
+  GONE: [
+    "A requested file or directory could not be found at the time an operation was processed.",
+    "NotFoundError",
+  ],
+  MISMATCH: [
+    "The path supplied exists, but was not an entry of requested type.",
+    "TypeMismatchError",
+  ],
+  MOD_ERR: [
+    "The object can not be modified in this way.",
+    "InvalidModificationError",
+  ],
+  SYNTAX: (m) => [
+    `Failed to execute 'write' on 'UnderlyingSinkBase': Invalid params passed. ${m}`,
+    "SyntaxError",
+  ],
+  SECURITY: [
+    "It was determined that certain files are unsafe for access within a Web application, or that too many calls are being made on file resources.",
+    "SecurityError",
+  ],
+  DISALLOWED: [
+    "The request is not allowed by the user agent or the platform in the current context.",
+    "NotAllowedError",
+  ],
+};
 
 const {
   ObjectPrototypeIsPrototypeOf,
@@ -30,8 +60,17 @@ class StorageManager {
   }
 
   // https://developer.mozilla.org/en-US/docs/Web/API/StorageManager/getDirectory
-  async getDirectory(name) {
-    return new FileSystemDirectoryHandle(name);
+  async getDirectory() {
+    const location = new URL(".", op_get_location());
+
+    if (location.protocol !== "file:") {
+      throw new DOMException(
+        "The getDirectory method is only available for file URLs.",
+        "SecurityError",
+      );
+    }
+
+    return new FileSystemDirectoryHandle(location.href, "");
   }
 
   [SymbolFor("Deno.privateCustomInspect")](inspect, inspectOptions) {
@@ -50,10 +89,12 @@ const storage = webidl.createBranded(StorageManager);
 
 class FileSystemHandle {
   #kind;
+  #root;
   #name;
 
-  constructor(kind, name) {
+  constructor(kind, root, name) {
     this.#kind = kind;
+    this.#root = root;
     this.#name = name;
   }
 
@@ -65,68 +106,113 @@ class FileSystemHandle {
     return this.#name;
   }
 
+  get ref() {
+    return join(this.#root, this.#name);
+  }
+
   async isSameEntry(other) {
     if (this === other) {
       return true;
     }
+
     return other instanceof FileSystemHandle &&
       this.kind == other.kind &&
-      this.name == other.name;
+      this.ref == other.ref;
   }
 }
 
 class FileSystemDirectoryHandle extends FileSystemHandle {
-  constructor(name) {
-    super("directory", name);
+  constructor(root, name) {
+    super("directory", root, name);
   }
 
   // https://developer.mozilla.org/en-US/docs/Web/API/FileSystemDirectoryHandle/getDirectoryHandle
   async getDirectoryHandle(name, options = {}) {
-    if (name === "") {
-      throw new TypeError(`Name can't be an empty string.`);
-    }
-    if (name === "." || name === ".." || name.includes("/")) {
-      throw new TypeError(`Name contains invalid characters.`);
-    }
+    checkName(name);
     options.create = !!options.create;
 
-    return new FileSystemDirectoryHandle(join(this.name, name));
+    const url = new URL(join(this.ref, name));
+    const stat = await fs.lstat(url).catch(() => {});
+    const isDirectory = stat?.isDirectory;
+    if (isDirectory) {
+      return new FileSystemDirectoryHandle(this.ref, name);
+    }
+    if (isDirectory === false) {
+      throw new DOMException(...errors.MISMATCH);
+    }
+    if (!options.create) {
+      throw new DOMException(...errors.GONE);
+    }
+    await fs.mkdir(url);
+    return new FileSystemDirectoryHandle(this.ref, name);
   }
 
   // https://developer.mozilla.org/en-US/docs/Web/API/FileSystemDirectoryHandle/getFileHandle
   async getFileHandle(name, options = {}) {
-    if (name === "") {
-      throw new TypeError(`Name can't be an empty string.`);
-    }
-    if (name === "." || name === ".." || name.includes("/")) {
-      throw new TypeError(`Name contains invalid characters.`);
-    }
+    checkName(name);
     options.create = !!options.create;
 
-    return new FileSystemFileHandle(join(this.name, name));
+    const url = new URL(join(this.ref, name));
+    const stat = await fs.lstat(url).catch(() => {});
+    const isFile = stat?.isFile;
+    if (isFile) {
+      return new FileSystemFileHandle(this.ref, name);
+    }
+    if (isFile === false) {
+      throw new DOMException(...errors.MISMATCH);
+    }
+    if (!options.create) {
+      throw new DOMException(...errors.GONE);
+    }
+    const c = await fs.open(url, { create: true, write: true });
+    c.close();
+    return new FileSystemFileHandle(this.ref, name);
   }
 
   // https://developer.mozilla.org/en-US/docs/Web/API/FileSystemDirectoryHandle/entries
   async *entries() {
-    for await (const entry of fs.readDir(new URL(this.name))) {
+    for await (const entry of fs.readDir(new URL(this.ref))) {
       if (entry.isDirectory) {
-        yield [entry.name, this.getDirectoryHandle(entry.name)];
+        yield [entry.name, new FileSystemDirectoryHandle(this.ref, entry.name)];
       } else if (entry.isFile) {
-        yield [entry.name, this.getFileHandle(entry.name)];
+        yield [entry.name, new FileSystemFileHandle(this.ref, entry.name)];
       }
     }
   }
 
   // https://developer.mozilla.org/en-US/docs/Web/API/FileSystemDirectoryHandle/removeEntry
   async removeEntry(name, options = {}) {
-    throw new Error("Not implemented");
-    // if (name === '') {
-    //   throw new TypeError(`Name can't be an empty string.`)
-    // }
-    // if (name === '.' || name === '..' || name.includes('/')) {
-    //   throw new TypeError(`Name contains invalid characters.`)
-    // }
-    // options.recursive = !!options.recursive // cuz node's fs.rm require boolean
+    checkName(name);
+    const url = new URL(join(this.ref, name));
+    try {
+      const stat = await fs.lstat(url);
+
+      if (stat.isDirectory) {
+        if (options.recursive) {
+          try {
+            await fs.remove(path, { recursive: true });
+          } catch (err) {
+            if (err.code === "ENOTEMPTY") {
+              throw new DOMException(...errors.MOD_ERR);
+            }
+            throw err;
+          }
+        } else {
+          try {
+            await fs.remove(path);
+          } catch (err) {
+            throw new DOMException(...MOD_ERR);
+          }
+        }
+      } else {
+        await fs.remove(path);
+      }
+    } catch (err) {
+      if (err.name === "NotFound") {
+        throw new DOMException(...errors.GONE);
+      }
+      throw err;
+    }
   }
 
   // https://developer.mozilla.org/en-US/docs/Web/API/FileSystemDirectoryHandle/resolve
@@ -187,22 +273,45 @@ Object.defineProperties(FileSystemDirectoryHandle.prototype, {
 });
 
 class FileSystemFileHandle extends FileSystemHandle {
-  constructor(name) {
-    super("file", name);
+  constructor(root, name) {
+    super("file", root, name);
   }
 
   // https://developer.mozilla.org/en-US/docs/Web/API/FileSystemFileHandle/createWritable
   async createWritable(options = {}) {
-    throw new Error("Not implemented");
-    // return new FileSystemWritableFileStream(
-    //   await this[kAdapter].createWritable(options)
-    // )
+    try {
+      const url = new URL(this.ref);
+      const fileHandle = await fs.open(url, {
+        write: true,
+        truncate: !options.keepExistingData,
+      });
+      const { size } = await fileHandle.stat();
+      return new FileSystemWritableFileStream(new Sink(fileHandle, size));
+    } catch (err) {
+      if (err.name === "NotFound") {
+        throw new DOMException(...errors.GONE);
+      }
+      throw err;
+    }
   }
 
   // https://developer.mozilla.org/en-US/docs/Web/API/FileSystemFileHandle/getFile
   async getFile() {
-    throw new Error("Not implemented");
-    // return new File(await this[kAdapter].getFile())
+    try {
+      const url = new URL(this.ref);
+      const [stats, content] = await Promise.all([
+        fs.stat(url),
+        fs.readFile(url),
+      ]);
+      return new File([content], this.name, {
+        lastModified: Number(stats.mtime),
+      });
+    } catch (err) {
+      if (err.name === "NotFound") {
+        throw new DOMException(...errors.GONE);
+      }
+      throw err;
+    }
   }
 }
 
@@ -219,23 +328,42 @@ Object.defineProperties(FileSystemFileHandle.prototype, {
 });
 
 class FileSystemWritableFileStream extends WritableStream {
+  #closed = false;
+
+  constructor(writer) {
+    super(writer);
+  }
+
   async close() {
-    throw new Error("Not implemented");
+    this.#closed = true;
+    const writer = this.getWriter();
+    const p = writer.close();
+    writer.releaseLock();
+    return p;
   }
 
   // https://developer.mozilla.org/en-US/docs/Web/API/FileSystemWritableFileStream/seek
   seek(position) {
-    throw new Error("Not implemented");
+    return this.write({ type: "seek", position });
   }
 
   // https://developer.mozilla.org/en-US/docs/Web/API/FileSystemWritableFileStream/truncate
   truncate(size) {
-    throw new Error("Not implemented");
+    return this.write({ type: "truncate", size });
   }
 
   // https://developer.mozilla.org/en-US/docs/Web/API/FileSystemWritableFileStream/write
   write(data) {
-    throw new Error("Not implemented");
+    if (this.#closed) {
+      return Promise.reject(
+        new TypeError("Cannot write to a CLOSED writable stream"),
+      );
+    }
+
+    const writer = this.getWriter();
+    const result = writer.write(data);
+    writer.releaseLock();
+    return result;
   }
 }
 
@@ -257,6 +385,84 @@ Object.defineProperties(FileSystemWritableFileStream.prototype, {
   write: { enumerable: true },
 });
 
+export class Sink {
+  constructor(fileHandle, size) {
+    this.fileHandle = fileHandle;
+    this.size = size;
+    this.position = 0;
+  }
+
+  async abort() {
+    await this.fileHandle.close();
+  }
+
+  async write(chunk) {
+    if (typeof chunk === "object") {
+      if (chunk.type === "write") {
+        if (Number.isInteger(chunk.position) && chunk.position >= 0) {
+          this.position = chunk.position;
+        }
+        if (!("data" in chunk)) {
+          await this.fileHandle.close();
+          throw new DOMException(
+            ...errors.SYNTAX("write requires a data argument"),
+          );
+        }
+        chunk = chunk.data;
+      } else if (chunk.type === "seek") {
+        if (Number.isInteger(chunk.position) && chunk.position >= 0) {
+          if (this.size < chunk.position) {
+            throw new DOMException(...errors.INVALID);
+          }
+          this.position = chunk.position;
+          return;
+        } else {
+          await this.fileHandle.close();
+          throw new DOMException(
+            ...errors.SYNTAX("seek requires a position argument"),
+          );
+        }
+      } else if (chunk.type === "truncate") {
+        if (Number.isInteger(chunk.size) && chunk.size >= 0) {
+          await this.fileHandle.truncate(chunk.size);
+          this.size = chunk.size;
+          if (this.position > this.size) {
+            this.position = this.size;
+          }
+          return;
+        } else {
+          await this.fileHandle.close();
+          throw new DOMException(
+            ...errors.SYNTAX("truncate requires a size argument"),
+          );
+        }
+      }
+    }
+
+    if (chunk instanceof ArrayBuffer) {
+      chunk = new Uint8Array(chunk);
+    } else if (typeof chunk === "string") {
+      chunk = new TextEncoder().encode(chunk);
+    } else if (chunk instanceof Blob) {
+      await this.fileHandle.seek(this.position, SeekMode.Start);
+      for await (const data of chunk.stream()) {
+        const written = await this.fileHandle.write(data);
+        this.position += written;
+        this.size += written;
+      }
+      return;
+    }
+    await this.fileHandle.seek(this.position, SeekMode.Start);
+    const written = await this.fileHandle.write(chunk);
+    this.position += written;
+    this.size += written;
+  }
+
+  async close() {
+    await this.fileHandle.close();
+  }
+}
+
 export {
   FileSystemDirectoryHandle,
   FileSystemFileHandle,
@@ -270,4 +476,13 @@ function join(root, name) {
     return root + name;
   }
   return root + "/" + name;
+}
+
+function checkName(name) {
+  if (name === "") {
+    throw new TypeError(`Name can't be an empty string.`);
+  }
+  if (name === "." || name === ".." || name.includes("/")) {
+    throw new TypeError(`Name contains invalid characters.`);
+  }
 }