+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
-## [0.1.1] - Unreleased
+## [0.1.1] - 2025-07-12
 ### Added
 - `location`
 - File System API
@@ -20,5 +20,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ## [0.1.0] - 2025-07-08
 First version
 
-[0.1.1]: https://github.com/oscarotero/toxojs/compare/v0.1.0...HEAD
+[0.1.1]: https://github.com/oscarotero/toxojs/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/oscarotero/toxojs/releases/tag/v0.1.0